requirements.txt
# Core Dagster
dagster>=1.8.0
dagster-webserver>=1.8.0
dagster-postgres>=0.23.0
dagster-duckdb>=0.23.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
duckdb>=0.9.0

# API & Web
requests>=2.31.0
httpx>=0.25.0

# Data validation# Core Dagster
dagster>=1.8.0
dagster-webserver>=1.8.0
dagster-postgres>=0.23.0
dagster-duckdb>=0.23.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
duckdb>=0.9.0

# API & Web
requests>=2.31.0
httpx>=0.25.0

# Data validation
great-expectations>=0.18.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
faker>=20.0.0
great-expectations>=0.18.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
faker>=20.0.0